image-service/requirements.txt
rembg[cpu]
fastapi
uvicorn
pillow
python-multipart
numpy
psutil
